Refactorizacion de usuario
Refactorizacion:
- en el login
- en la obtención de la info del token
