+++ src/interfaces/variable.interface.ts
@@ -1,3 +1,5 @@
+import { IUsuario } from './usuario.interface';
+
 //Listado de variables que queremos usar en el Servicio de resolver-operations
 export interface IVariables{    
     // Categoria
@@ -20,4 +22,6 @@ export interface IVariables{
     usuario_modificacion?: string;
 
     nuevoRegistro?: IVariables;
+
+    usuario?: IUsuario;
 }
